+++ lib/utils.dart
@@ -2,6 +2,7 @@ library elec_server.utils;
 
 export 'src/utils/parse_custom_integer_range.dart';
 export 'src/utils/file_utils.dart';
+export 'src/utils/iterable_extensions.dart';
 export 'src/utils/layout_tree.dart';
 export 'src/utils/lib_plotly.dart';
 export 'src/utils/term_cache.dart';